+++ src/hxCUtils.c
@@ -15,16 +15,16 @@ struct hxSettings g_hxSettings;
 #if defined(__clang__)
 __attribute__((no_sanitize("address")))
 #endif
-void hxHexDump(const void* address, uint32_t bytes, int pretty) {
+void hxHexDump(const void* address, size_t bytes, int pretty) {
 	if ((HX_RELEASE) < 2 && address != hxnull) {
-		bytes = (bytes + 15u) & ~(uint32_t)15; // round up to 16 bytes.
+		bytes = (bytes + 15u) & ~(size_t)15; // round up to 16 bytes.
 		const uint8_t* addr = (const uint8_t*)address;
-		for (uint32_t i = 0; i < bytes;) {
+		for (size_t i = 0; i < bytes;) {
 			if (pretty) {
 				hxLogConsole("%08x: ", (unsigned int)(uintptr_t)addr);
 			}
 			const uint8_t* str = addr;
-			for (int32_t maximum = 4; i < bytes && maximum--; i += 4) {
+			for (size_t maximum = 4; i < bytes && maximum--; i += 4) {
 				hxLogConsole("%02x%02x%02x%02x ", addr[0], addr[1], addr[2], addr[3]);
 				addr += 4;
 			}
@@ -42,11 +42,11 @@ void hxHexDump(const void* address, uint32_t bytes, int pretty) {
 #if defined(__clang__)
 __attribute__((no_sanitize("address")))
 #endif
-void hxFloatDump(const float* address, uint32_t count) {
+void hxFloatDump(const float* address, size_t count) {
 	if ((HX_RELEASE) < 2 && address != hxnull) {
-		for (uint32_t i = 0; i < count;) {
+		for (size_t i = 0; i < count;) {
 			hxLogConsole("%08x: ", (unsigned int)(uintptr_t)address);
-			for (int32_t maximum = 4; i < count && maximum--; i++) {
+			for (size_t maximum = 4; i < count && maximum--; i++) {
 				hxLogConsole("%8f ", *address++);
 			}
 			hxLogConsole("\n");
