+++ src/hxProfiler.cpp
@@ -53,10 +53,10 @@ void hxProfiler::log() {
 	HX_PROFILER_LOCK();
 	m_isStarted = false;
 
-	for (uint32_t i = 0; i < m_records.size(); ++i) {
+	for (size_t i = 0; i < m_records.size(); ++i) {
 		const hxProfiler::Record& rec = m_records[i];
 
-		uint32_t delta = rec.m_end - rec.m_begin;
+		size_t delta = rec.m_end - rec.m_begin;
 		hxLogRelease("profile %s: %fms cycles %u thread %x\n", hxBasename(rec.m_label),
 			delta * (double)c_hxTimeMillisecondsPerCycle, (unsigned int)delta,
 			(unsigned int)rec.m_threadId);
@@ -71,8 +71,8 @@ void hxProfiler::writeToChromeTracing(const char* filename) {
 
 	f.print("[\n");
 	// Converting absolute values works better with integer precision.
-	uint32_t cyclesPerMicrosecond = (uint32_t)(1.0e-3f / c_hxTimeMillisecondsPerCycle);
-	for (uint32_t i = 0; i < m_records.size(); ++i) {
+	size_t cyclesPerMicrosecond = (size_t)(1.0e-3f / c_hxTimeMillisecondsPerCycle);
+	for (size_t i = 0; i < m_records.size(); ++i) {
 		const hxProfiler::Record& rec = m_records[i];
 		if (i != 0) { f.print(",\n"); }
 		const char* bn = hxBasename(rec.m_label);
